Cambio el tipo del id de las entidades de Long a Integer porque la base no soporta el Long.

+++ trunk/gestionTurnos/src/main/java/ar/com/trivoli/gestionturnos/obrasocial/model/ObraSocial.java
@@ -17,7 +17,7 @@ import javax.persistence.Table;
 
 @Entity
 @Table(name = "obrasSociales")
-public class ObraSocial extends EntidadBase<Long> {
+public class ObraSocial extends EntidadBase<Integer> {
 
 	/**
 	 *
